+++ web/modules/custom/hoeringsportal_citizen_proposal/src/Form/ProposalAdminForm.php
@@ -4,20 +4,20 @@ namespace Drupal\hoeringsportal_citizen_proposal\Form;
 
 use Drupal\Core\Form\FormBase;
 use Drupal\Core\Form\FormStateInterface;
-use Drupal\Core\State\State;
+use Drupal\hoeringsportal_citizen_proposal\Helper\Helper;
+use Drupal\hoeringsportal_citizen_proposal\Helper\MailHelper;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Form for adding proposal.
  */
 final class ProposalAdminForm extends FormBase {
-  public const ADMIN_FORM_VALUES_STATE_KEY = 'citizen_proposal_admin_form_values';
 
   /**
    * Constructor for the proposal add form.
    */
   public function __construct(
-    readonly private State $state,
+    readonly private Helper $helper
   ) {
   }
 
@@ -26,7 +26,7 @@ final class ProposalAdminForm extends FormBase {
    */
   public static function create(ContainerInterface $container) {
     return new static(
-      $container->get('state'),
+      $container->get(Helper::class),
     );
   }
 
@@ -41,7 +41,7 @@ final class ProposalAdminForm extends FormBase {
    * {@inheritdoc}
    */
   public function buildForm(array $form, FormStateInterface $form_state) {
-    $adminFormStateValues = $this->state->get(self::ADMIN_FORM_VALUES_STATE_KEY);
+    $adminFormStateValues = $this->helper->getAdminValue();
 
     $form['authenticate'] = [
       '#type' => 'details',
@@ -230,6 +230,8 @@ final class ProposalAdminForm extends FormBase {
       '#default_value' => $adminFormStateValues['sidebar_text']['value'] ?? '',
     ];
 
+    $this->buildEmailsForm($form, $adminFormStateValues ?? []);
+
     $form['actions']['#type'] = 'actions';
     $form['actions']['submit'] = [
       '#type' => 'submit',
@@ -241,10 +243,69 @@ final class ProposalAdminForm extends FormBase {
   }
 
   /**
+   * Build emails form.
+   *
+   * @param array $form
+   *   The form.
+   * @param array $adminFormStateValues
+   *   The admin form state values.
+   *
+   * @return array
+   *   The form.
+   */
+  private function buildEmailsForm(array &$form, array $adminFormStateValues): array {
+    $form['emails'] = [
+      '#type' => 'details',
+      '#tree' => TRUE,
+      '#open' => TRUE,
+      '#title' => $this
+        ->t('Emails'),
+    ];
+
+    $form['emails']['description'] = [
+      '#markup' => $this->t('You can use tokens in email subject and both tokens and Twig in email content.'),
+    ];
+
+    $form['emails']['email_editor'] = [
+      '#type' => 'email',
+      '#title' => $this->t('Editor email address'),
+      '#required' => TRUE,
+      '#default_value' => $adminFormStateValues['emails']['email_editor'] ?? '',
+    ];
+
+    foreach ([
+      MailHelper::MAILER_SUBTYPE_PROPOSAL_CREATED_CITIZEN => $this->t('Proposal created (citizen)'),
+      MailHelper::MAILER_SUBTYPE_PROPOSAL_CREATED_EDITOR => $this->t('Proposal created (editor)'),
+      MailHelper::MAILER_SUBTYPE_PROPOSAL_PUBLISHED_CITIZEN => $this->t('Proposal published (citizen)'),
+    ] as $key => $title) {
+      $form['emails'][$key] = [
+        '#type' => 'fieldset',
+        '#title' => $title,
+
+        'subject' => [
+          '#type' => 'textfield',
+          '#title' => $this->t('Subject'),
+          '#required' => TRUE,
+          '#default_value' => $adminFormStateValues['emails'][$key]['subject'] ?? '',
+        ],
+        'content' => [
+          '#type' => 'text_format',
+          '#title' => $this->t('Content'),
+          '#required' => TRUE,
+          '#format' => $adminFormStateValues['emails'][$key]['content']['format'] ?? 'email_html',
+          '#default_value' => $adminFormStateValues['emails'][$key]['content']['value'] ?? '',
+        ],
+      ];
+    }
+
+    return $form;
+  }
+
+  /**
    * {@inheritdoc}
    */
   public function submitForm(array &$form, FormStateInterface $formState): void {
-    $this->state->set(self::ADMIN_FORM_VALUES_STATE_KEY, $formState->getValues());
+    $this->helper->setAdminValues($formState->getValues());
   }
 
 }