+++ CHANGELOG.md
@@ -6,12 +6,16 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+* [PR-337](https://github.com/itk-dev/hoeringsportal/pull/337)
+  Updated drupal core 9.5.10 and contrib modules
 * [PR-332](https://github.com/itk-dev/hoeringsportal/pull/332)
   Added data storage consent checkbox on citizen proposal form
 * [PR-333](https://github.com/itk-dev/hoeringsportal/pull/333)
   Removed email from citizen proposal support form
 * [PR-335](https://github.com/itk-dev/hoeringsportal/pull/335)
   Hid “Author email display” from display
+* [PR-333](https://github.com/itk-dev/hoeringsportal/pull/333)
+  Sent notification mails on citizen proposal creation and publication
 
 ## [3.1.0] - 2023-08-04 - Citizen proposal